[NodeDescription] Factor out NodeDescription
Move NodeDescriptions to a seperate file. Keeping NodeDescriptions in a
separate file is cleaner. Moreover, it is important if IPTopo requires
the import of NodeDescriptions. This avoids forward references, when
IPTopo should stay at the top of `iptopo.py`.
