+++ ipmininet/iptopo.py
@@ -1,5 +1,4 @@
 """This module defines topology class that supports adding L3 routers"""
-import functools
 import itertools
 from typing import Union, Type, Dict, List, Optional, Tuple, Any
 
@@ -8,10 +7,13 @@ from mininet.log import lg
 
 from ipmininet.overlay import Overlay, Subnet
 from ipmininet.utils import get_set, is_container
+from ipmininet.node_description import NodeDescription, RouterDescription,\
+    HostDescription, LinkDescription
+from ipmininet.router import Router
 from ipmininet.router.config import BasicRouterConfig, OSPFArea, AS,\
     iBGPFullMesh, OpenrDomain
-from ipmininet.router.config.base import Daemon, RouterConfig, NodeConfig
-from ipmininet.host.config import HostConfig, DNSZone
+from ipmininet.router.config.base import Daemon, NodeConfig
+from ipmininet.host.config import DNSZone
 from ipmininet.ipnet import IPNet
 
 
@@ -58,11 +60,17 @@ class IPTopo(Topo):
            :param name: the name of the node"""
         return HostDescription(super().addHost(str(name), **kwargs), self)
 
-    def addRouter(self, name: str, **kwargs) -> 'RouterDescription':
+    def addRouter(self,
+                  name: str,
+                  routerDescription: 'RouterDescription' = RouterDescription,
+                  **kwargs) -> 'RouterDescription':
         """Add a router to the topology
 
-        :param name: the name of the node"""
-        return RouterDescription(self.addNode(str(name), isRouter=True,
+        :param name: the name of the node
+        "param routerDescription: the RouterDescription class to return
+            (optional)"""
+        return routerDescription(self.addNode(str(name),
+                                              isRouter=True,
                                               **kwargs), self)
 
     def addRouters(self, *routers: Union[str, Tuple[str, Dict[str, Any]]],
@@ -242,117 +250,3 @@ class OverlayWrapper:
 
     def __call__(self, *args, **kwargs):
         return self.topo.addOverlay(self.overlay(*args, **kwargs))
-
-
-class NodeDescription(str):
-
-    def __new__(cls, value, *args, **kwargs):
-        return super().__new__(cls, value)
-
-    def __init__(self, o, topo: Optional[IPTopo] = None):
-        self.topo = topo
-        super().__init__()
-
-    def addDaemon(self, daemon: Union[Daemon, Type[Daemon]],
-                  default_cfg_class: Type[NodeConfig] = BasicRouterConfig,
-                  cfg_daemon_list="daemons", **daemon_params):
-        """Add the daemon to the list of daemons to start on the node.
-
-        :param daemon: daemon class
-        :param default_cfg_class: config class to use
-            if there is no configuration class defined for the router yet.
-        :param cfg_daemon_list: name of the parameter containing
-            the list of daemons in your config class constructor.
-            For instance, RouterConfig uses 'daemons'
-            but BasicRouterConfig uses 'additional_daemons'.
-        :param daemon_params: all the parameters to give
-            when instantiating the daemon class."""
-        if self.topo is None:
-            return
-        self.topo.addDaemon(self, daemon, default_cfg_class=default_cfg_class,
-                            cfg_daemon_list=cfg_daemon_list, **daemon_params)
-
-    def get_config(self, daemon: Union[Daemon, Type[Daemon]], **kwargs):
-        if self.topo is None:
-            return
-        return daemon.get_config(topo=self.topo, node=self, **kwargs)
-
-
-class RouterDescription(NodeDescription):
-    def addDaemon(self, daemon: Union[Daemon, Type[Daemon]],
-                  default_cfg_class: Type[RouterConfig] = BasicRouterConfig,
-                  **kwargs):
-        super(RouterDescription, self)\
-            .addDaemon(daemon, default_cfg_class=default_cfg_class, **kwargs)
-
-
-class HostDescription(NodeDescription):
-
-    def addDaemon(self, daemon: Union[Daemon, Type[Daemon]],
-                  default_cfg_class: Type[HostConfig] = HostConfig, **kwargs):
-        super(HostDescription, self)\
-            .addDaemon(daemon, default_cfg_class=default_cfg_class, **kwargs)
-
-
-@functools.total_ordering
-class LinkDescription:
-
-    def __init__(self, topo: IPTopo, src: str, dst: str, key, link_attrs: Dict):
-        self.src = src
-        self.dst = dst
-        self.key = key
-        self.link_attrs = link_attrs
-        self.src_intf = IntfDescription(self.src, topo, self,
-                                        self.link_attrs.setdefault("params1",
-                                                                   {}))
-        self.dst_intf = IntfDescription(self.dst, topo, self,
-                                        self.link_attrs.setdefault("params2",
-                                                                   {}))
-        super().__init__()
-
-    def __getitem__(self, item):
-        if isinstance(item, int):
-            if item == 0:
-                return self.src_intf
-            if item == 1:
-                return self.dst_intf
-            if item == 3:
-                return self.key
-            raise IndexError("Links have only two nodes and one key")
-
-        if item == self.src:
-            return self.src_intf
-        if item == self.dst:
-            return self.dst_intf
-        raise KeyError("Node '%s' is not on this link" % item)
-
-    # The following methods allow this object to behave like an edge key
-    # for mininet.topo.MultiGraph
-
-    def __hash__(self):
-        return self.key.__hash__()
-
-    def __eq__(self, other):
-        return self.key == other
-
-    def __lt__(self, other):
-        return self.key.__lt__(other)
-
-
-class IntfDescription(NodeDescription):
-
-    def __init__(self, o: str, topo: IPTopo, link: LinkDescription,
-                 intf_attrs: Dict):
-        self.link = link
-        self.node = o
-        self.intf_attrs = intf_attrs
-        super().__init__(o, topo)
-
-    def addParams(self, **kwargs):
-        self.intf_attrs.update(kwargs)
-
-    def __hash__(self):
-        return self.node.__hash__()
-
-    def __eq__(self, other):
-        return self.node.__eq__(other)
